requirements.txt
vllm>=0.6.0
pyyaml>=6.0
matplotlib>=3.8
numpy>=2.0
networkx>=3.2
plotly>=5.18
